+++ Adapters/IronSource/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## IronSource iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationIronSource', '9.1.0.0'
+pod 'CASMediationIronSource', '9.2.0.0'
 ```
+
+### 9.2.0.0
+- Certified with IronSource - 9.2.0
 
 ### 9.1.0.0
 - Certified with IronSource - 9.1.0