+++ Adapters/YangoAds/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Yango Ads iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationYangoAds', '7.18.0.0'
+pod 'CASMediationYangoAds', '7.18.1.0'
 ```
+
+### 7.18.1.0
+- Certified with YangoAds - 7.18.1
+- Fixed a rare issue where initialization would end with a timeout.
 
 ### 7.18.0.0
 - Certified with YangoAds - 7.18.0